Gets rid of NPE when uninstalling
If the package contains a non-OSGi jar file, the uninstaller would get an NPE.

Fixed both at the point where the NPE would happen, as well preventing the collection of symbolic names from having nulls in it in the first place. Also added logging that will warn the user that the package contains non-OSGi jars, which shouldn't happen.

Fixes GH-9